Replaced int xprts_nr with size_t xprts_nr

+++ colibri/colibri_setup.h
@@ -134,7 +134,7 @@ struct c2_colibri {
 	/**
 	   Size of cc_xprts array.
 	 */
-	int                       cc_xprts_nr;
+	size_t                    cc_xprts_nr;
 
         /**
            List of network domain per colibri context.
@@ -170,7 +170,7 @@ struct c2_colibri {
 	 * @see c2_net_transfer_mc:ntm_recv_queue_length
 	 * Default is set to C2_NET_TM_RECV_QUEUE_DEF_LEN.
 	 */
-	uint32_t                 cc_recv_queue_min_length;
+	uint32_t                  cc_recv_queue_min_length;
 
 	/** Maximum RPC message size. */
 	uint32_t                  cc_max_rpc_msg_size;
@@ -187,7 +187,7 @@ struct c2_colibri {
    @param out File descriptor to which output is written
  */
 int c2_cs_init(struct c2_colibri *cs_colibri, struct c2_net_xprt **xprts,
-	       int xprts_nr, FILE *out);
+	       size_t xprts_nr, FILE *out);
 
 /**
    Finalises colibri context.
